+++ src/Module/GenerateSessionsHandler.php
@@ -2,28 +2,41 @@
 
 namespace RebelCode\EddBookings\Sessions\Module;
 
+use AppendIterator;
+use ArrayAccess;
+use ArrayIterator;
 use Dhii\Data\Container\ContainerGetCapableTrait;
 use Dhii\Data\Container\ContainerGetPathCapableTrait;
 use Dhii\Data\Container\CreateContainerExceptionCapableTrait;
 use Dhii\Data\Container\CreateNotFoundExceptionCapableTrait;
-use Dhii\Data\Container\DeleteCapableInterface;
 use Dhii\Data\Container\NormalizeKeyCapableTrait;
 use Dhii\Exception\CreateInvalidArgumentExceptionCapableTrait;
 use Dhii\Exception\CreateOutOfRangeExceptionCapableTrait;
 use Dhii\Factory\FactoryInterface;
 use Dhii\I18n\StringTranslatingTrait;
 use Dhii\Invocation\InvocableInterface;
+use Dhii\Iterator\NormalizeIteratorCapableTrait;
+use Dhii\Storage\Resource\DeleteCapableInterface;
 use Dhii\Storage\Resource\InsertCapableInterface;
-use Dhii\Storage\Resource\SelectCapableInterface;
-use Dhii\Time\PeriodInterface;
+use Dhii\Util\Normalization\NormalizeArrayCapableTrait;
 use Dhii\Util\Normalization\NormalizeIntCapableTrait;
 use Dhii\Util\Normalization\NormalizeIterableCapableTrait;
 use Dhii\Util\Normalization\NormalizeStringCapableTrait;
-use Dhii\Util\String\StringableInterface as Stringable;
+use IteratorIterator;
+use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use Psr\EventManager\EventInterface;
-use RebelCode\Sessions\SessionGeneratorInterface;
+use RebelCode\Bookings\Availability\AvailabilityInterface;
+use RebelCode\Bookings\Availability\AvailabilityPeriodInterface;
+use RebelCode\Bookings\Availability\CompositeAvailability;
+use RebelCode\Bookings\Availability\IntersectionAvailability;
+use RebelCode\Bookings\Sessions\SessionGeneratorInterface;
+use RebelCode\EddBookings\Sessions\Time\Period;
+use RebelCode\EddBookings\Sessions\Util\ModifyCallbackIterator;
+use RebelCode\Entity\EntityManagerInterface;
+use RebelCode\Entity\GetCapableManagerInterface;
 use RebelCode\Time\NormalizeTimestampCapableTrait;
+use stdClass;
 use Traversable;
 
 /**
@@ -46,6 +59,9 @@ class GenerateSessionsHandler implements InvocableInterface
     use NormalizeIterableCapableTrait;
 
     /* @since [*next-version*] */
+    use NormalizeIteratorCapableTrait;
+
+    /* @since [*next-version*] */
     use NormalizeIntCapableTrait;
 
     /* @since [*next-version*] */
@@ -53,6 +69,9 @@ class GenerateSessionsHandler implements InvocableInterface
 
     /* @since [*next-version*] */
     use NormalizeStringCapableTrait;
+
+    /* @since [*next-version*] */
+    use NormalizeArrayCapableTrait;
 
     /* @since [*next-version*] */
     use CreateContainerExceptionCapableTrait;
@@ -70,13 +89,49 @@ class GenerateSessionsHandler implements InvocableInterface
     use StringTranslatingTrait;
 
     /**
-     * The SELECT RM for session generator rule.
+     * The session generator.
      *
      * @since [*next-version*]
      *
-     * @var SelectCapableInterface
+     * @var SessionGeneratorInterface
      */
-    protected $rulesSelectRm;
+    protected $generator;
+
+    /**
+     * The services entity manager.
+     *
+     * @since [*next-version*]
+     *
+     * @var EntityManagerInterface
+     */
+    protected $servicesManager;
+
+    /**
+     * The resources entity manager.
+     *
+     * @since [*next-version*]
+     *
+     * @var GetCapableManagerInterface
+     */
+    protected $resourcesManager;
+
+    /**
+     * The factory for creating session types.
+     *
+     * @since [*next-version*]
+     *
+     * @var FactoryInterface
+     */
+    protected $sessionTypeFactory;
+
+    /**
+     * The factory for creating availabilities.
+     *
+     * @since [*next-version*]
+     *
+     * @var FactoryInterface
+     */
+    protected $availabilityFactory;
 
     /**
      * The INSERT RM for sessions.
@@ -106,49 +161,37 @@ class GenerateSessionsHandler implements InvocableInterface
     protected $exprBuilder;
 
     /**
-     * The session generator rule factory.
-     *
-     * @since [*next-version*]
-     *
-     * @var FactoryInterface
-     */
-    protected $ruleFactory;
-
-    /**
-     * The session generator factory.
-     *
-     * @since [*next-version*]
-     *
-     * @var FactoryInterface
-     */
-    protected $generatorFactory;
-
-    /**
      * Constructor.
      *
      * @since [*next-version*]
      *
-     * @param FactoryInterface       $generatorFactory The session generator factory.
-     * @param FactoryInterface       $ruleFactory      The session generator rule factory.
-     * @param SelectCapableInterface $rulesSelectRm    The SELECT RM for session generator rules.
-     * @param InsertCapableInterface $sessionsInsertRm The INSERT RM for sessions.
-     * @param DeleteCapableInterface $sessionsDeleteRm The DELETE RM for sessions.
-     * @param object                 $exprBuilder      The expression builder.
+     * @param SessionGeneratorInterface  $generator           The session generator.
+     * @param EntityManagerInterface     $servicesManager     The services entity manager.
+     * @param GetCapableManagerInterface $resourcesManager    The resources entity manager.
+     * @param FactoryInterface           $sessionTypeFactory  The factory for creating session types.
+     * @param FactoryInterface           $availabilityFactory The factory for creating availabilities.
+     * @param InsertCapableInterface     $sessionsInsertRm    The INSERT RM for sessions.
+     * @param DeleteCapableInterface     $sessionsDeleteRm    The DELETE RM for sessions.
+     * @param object                     $exprBuilder         The expression builder.
      */
     public function __construct(
-        $generatorFactory,
-        $ruleFactory,
-        $rulesSelectRm,
-        $sessionsInsertRm,
-        $sessionsDeleteRm,
+        SessionGeneratorInterface $generator,
+        EntityManagerInterface $servicesManager,
+        GetCapableManagerInterface $resourcesManager,
+        FactoryInterface $sessionTypeFactory,
+        FactoryInterface $availabilityFactory,
+        InsertCapableInterface $sessionsInsertRm,
+        DeleteCapableInterface $sessionsDeleteRm,
         $exprBuilder
     ) {
-        $this->generatorFactory = $generatorFactory;
-        $this->rulesSelectRm    = $rulesSelectRm;
-        $this->sessionsInsertRm = $sessionsInsertRm;
-        $this->sessionsDeleteRm = $sessionsDeleteRm;
-        $this->exprBuilder      = $exprBuilder;
-        $this->ruleFactory      = $ruleFactory;
+        $this->generator           = $generator;
+        $this->servicesManager     = $servicesManager;
+        $this->resourcesManager    = $resourcesManager;
+        $this->sessionTypeFactory  = $sessionTypeFactory;
+        $this->availabilityFactory = $availabilityFactory;
+        $this->sessionsInsertRm    = $sessionsInsertRm;
+        $this->sessionsDeleteRm    = $sessionsDeleteRm;
+        $this->exprBuilder         = $exprBuilder;
     }
 
     /**
@@ -166,113 +209,212 @@ class GenerateSessionsHandler implements InvocableInterface
             );
         }
 
-        $serviceId = $event->getParam('service_id');
+        $serviceId  = $event->getParam('service_id');
+        $resourceId = $event->getParam('resource_id');
 
-        if ($serviceId === null) {
+        // If neither a service nor a resource ID are given, stop
+        if ($serviceId === null && $resourceId === null) {
             return;
         }
 
-        $this->_generateForService($serviceId);
+        // Get the list of services.
+        // If a service ID is given, use only that service.
+        // If a resource ID is given, use the services that use that resource
+        $services = ($serviceId === null)
+            ? $this->_getServicesForResource($resourceId)
+            : [$this->servicesManager->get($serviceId)];
+
+        // Generate sessions for each applicable service
+        foreach ($services as $_service) {
+            $this->_generateForService($_service);
+        }
     }
 
     /**
-     * Generates sessions for a particular service, by ID.
+     * Generates sessions for a particular service.
      *
      * @since [*next-version*]
      *
-     * @param int|string|Stringable $serviceId The ID of the service for which to generate sessions.
+     * @param array|stdClass|ArrayAccess|ContainerInterface $service The service for which to generate sessions.
      */
-    protected function _generateForService($serviceId)
+    protected function _generateForService($service)
     {
-        $postId = func_get_arg(0);
+        // Get the service's schedule availability, since service availability it saved in its schedule
+        $scheduleId = $this->_containerGet($service, 'schedule_id');
+        $schedule   = $this->resourcesManager->get($scheduleId);
+        $scheduleAv = $this->_getResourceAvailability($schedule);
 
-        // Get the session types for the service
-        $sessionTypes = $this->_getPostMeta($postId, 'eddbk_session_types', []);
-        $durations    = [];
-        foreach ($sessionTypes as $_sessionType) {
+        // Create the resource availabilities
+        // The resources are retrieved from the service's session types
+        // Each session type has a list of resources associated with it
+        // Therefore, we need to iterate the session types, obtain their resources and get their availabilities
+        // Session type instances are also prepared during this iteration since they are required by the generator
+        // A cache of resources, keyed by their IDs, is also kept to minimize calls to the entity manager.
+        $resources        = [];
+        $resourceAvs      = [];
+        $sessionTypes     = [];
+        $sessionTypesData = $this->_containerGet($service, 'session_types');
+
+        foreach ($sessionTypesData as $_data) {
+            // Create and store session type instance
+            $_sessionType = $this->sessionTypeFactory->make($_data);
+
             try {
-                $durations[] = $this->_containerGetPath($_sessionType, ['data', 'duration']);
+                $_stResourceIds = $this->_containerGetPath($_data, ['data', 'resources']);
             } catch (NotFoundExceptionInterface $exception) {
                 continue;
             }
+
+            // This is the "ungrouped" behavior
+            // For the current version of EDD Bookings, the ungrouping of session types will happen here, during
+            // session generation. In the future, this will be handled by another component.
+            // $sessionTypes[] = [
+            //     'object'    => $_sessionType,
+            //     'resources' => [$_stResourceIds],
+            // ];
+
+            // Skip un-grouping of session types if it has no resources, otherwise the session type won't be stored
+            if (empty($_stResourceIds)) {
+                $sessionTypes[] = [
+                    'object'    => $_sessionType,
+                    'resources' => [],
+                ];
+                continue;
+            }
+
+            foreach ($_stResourceIds as $_resourceId) {
+                try {
+
+                    // Get and store the resource availability, unless it's already recorded
+                    if (!isset($resourceAvs[$_resourceId])) {
+                        // Get from cache first if available, otherwise get using the entity manager
+                        $_resource = !isset($resources[$_resourceId])
+                            ? $this->resourcesManager->get($_resourceId)
+                            : $resources[$_resourceId];
+
+                        $resourceAvs[$_resourceId] = $this->_getResourceAvailability($_resource);
+                    }
+
+                    // Temporary solution for ungrouping session types
+                    $sessionTypes[] = [
+                        'object'    => $_sessionType,
+                        'resources' => [$_resourceId],
+                    ];
+                } catch (NotFoundExceptionInterface $exception) {
+                    continue;
+                }
+            }
         }
 
-        $b = $this->exprBuilder;
-        // Get the session generator rules for the service
-        $rules = $this->rulesSelectRm->select(
-            $b->eq(
-                $b->ef('session_rule', 'service_id'),
-                $b->lit($postId)
-            )
-        );
+        // The final service availability is the intersection of the schedule availability with the composite
+        // availability of all the resources retrieved from the session types.
+        // The composition of resources yields all the available periods for all the resources.
+        // The intersection restricts those periods to those that are also present in the schedule
+        $availability = (count($resourceAvs) > 0)
+            ? new IntersectionAvailability([$scheduleAv, new CompositeAvailability($resourceAvs)])
+            : $scheduleAv;
 
+        // Use an append iterator to incrementally add more iterators, as retrieved from each generation pass
+        $sessions = new AppendIterator();
+        // Generate for 5 years by default
+        $range = new Period(time(), strtotime('+5 years'));
+
+        // Iterate all available periods of time
+        // For each period, check which session types need to be generated for it. This is determined by resource ID
+        // equivalence. If the availability period and the session types have the same resources, then the session
+        // type may be used to generate sessions for that period.
+        /* @var $_period AvailabilityPeriodInterface */
+        foreach ($availability->getAvailablePeriods($range) as $_period) {
+            $_sessionTypeObjects = [];
+
+            foreach ($sessionTypes as $_sessionType) {
+                $diff = array_diff($_sessionType['resources'], $_period->getResourceIds());
+
+                if (count($diff) === 0) {
+                    $_sessionTypeObjects[] = $_sessionType['object'];
+                }
+            }
+
+            $sessions->append(
+                $this->_normalizeIterator(
+                    $this->generator->generate($_period, $_sessionTypeObjects)
+                )
+            );
+        }
+
+        // Use a callback iterator to modify each session at the last minute to add the service ID
+        // The resource IDs also need to be imploded into a comma separated list string
+        $serviceId     = $this->_containerGet($service, 'id');
+        $finalSessions = new ModifyCallbackIterator($sessions, function ($session) use ($serviceId) {
+            $session['service_id']   = $serviceId;
+            $session['resource_ids'] = implode(',', $session['resource_ids']);
+
+            return $session;
+        });
+
+        // Delete all existing sessions for this service
+        $b = $this->exprBuilder;
         $this->sessionsDeleteRm->delete($b->eq(
             $b->var('service_id'),
             $b->lit($serviceId)
         ));
 
-        foreach ($rules as $_ruleCfg) {
-            $_rule   = $this->ruleFactory->make($_ruleCfg);
-            $_ruleId = $this->_containerGet($_ruleCfg, 'id');
-
-            // Initialize a generator with the durations
-            $generator = $this->generatorFactory->make([
-                'session_factory' => $this->_getSessionFactory($postId, $postId, $_ruleId),
-                'session_lengths' => $durations,
-            ]);
-
-            $this->_generateForRule($_rule, $generator);
-        }
+        // Save the newly generated session
+        $this->sessionsInsertRm->insert($finalSessions);
     }
 
     /**
-     * Generates sessions for a single rule, using a specific generator.
+     * Retrieve a resource's availability.
      *
      * @since [*next-version*]
      *
-     * @param Traversable               $rule      The rule as a traversable list of periods for each occurrence.
-     * @param SessionGeneratorInterface $generator The session generator instance to use.
+     * @param array|stdClass|ArrayAccess|ContainerInterface $resource The resource data.
+     *
+     * @return AvailabilityInterface The resource's full availability.
      */
-    protected function _generateForRule($rule, $generator)
+    protected function _getResourceAvailability($resource)
     {
-        foreach ($rule as $occurrence) {
-            /* @var $occurrence PeriodInterface */
-            $sessions = $generator->generate(
-                $this->_normalizeTimestamp($occurrence->getStart()),
-                $this->_normalizeTimestamp($occurrence->getEnd())
-            );
+        $availRules     = $this->_containerGetPath($resource, ['availability', 'rules']);
+        $timezone       = $this->_containerGetPath($resource, ['availability', 'timezone']);
+        $availabilities = [];
 
-            $this->sessionsInsertRm->insert($sessions);
+        foreach ($availRules as $_ruleData) {
+            $_config                 = $this->_normalizeArray($_ruleData);
+            $_config['timezone']     = $timezone;
+            $_config['resource_ids'] = [$_config['resource_id']];
+            $availabilities[]        = $this->availabilityFactory->make($_config);
         }
+
+        return new CompositeAvailability($availabilities);
     }
 
     /**
-     * Retrieves the session factory to use when generating sessions for a given service, resource and rule.
+     * Retrieves the services that use a specific resource, by ID.
      *
      * @since [*next-version*]
      *
-     * @param int|string|Stringable $serviceId  The service ID.
-     * @param int|string|Stringable $resourceId The resource ID.
-     * @param int|string|Stringable $ruleId     The session rule ID.
+     * @param int|string $resourceId The resource ID.
      *
-     * @return callable The session factory, as a callable that receives the session start and end timestamps as
-     *                  arguments and returns a session.
+     * @return array|stdClass|Traversable A list of service containers.
      */
-    protected function _getSessionFactory($serviceId, $resourceId, $ruleId)
+    protected function _getServicesForResource($resourceId)
     {
-        $serviceId  = $this->_normalizeInt($serviceId);
-        $resourceId = $this->_normalizeInt($resourceId);
-        $ruleId     = $this->_normalizeInt($ruleId);
+        $results  = [];
+        $services = $this->servicesManager->query();
 
-        return function ($start, $end) use ($serviceId, $resourceId, $ruleId) {
-            return [
-                'start'       => $start,
-                'end'         => $end,
-                'service_id'  => $serviceId,
-                'resource_id' => $resourceId,
-                'rule_id'     => $ruleId,
-            ];
-        };
+        foreach ($services as $_service) {
+            $_sessionTypes = $this->_containerGet($_service, 'session_types');
+
+            foreach ($_sessionTypes as $_sessionType) {
+                $_resources = $this->_containerGetPath($_sessionType, ['data', 'resources']);
+
+                if (in_array($resourceId, $_resources)) {
+                    $results[] = $_service;
+                }
+            }
+        }
+
+        return $results;
     }
 
     /**
@@ -293,5 +435,25 @@ class GenerateSessionsHandler implements InvocableInterface
         return ($metaValue === '')
             ? $default
             : $metaValue;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _createArrayIterator(array $array)
+    {
+        return new ArrayIterator($array);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _createTraversableIterator(Traversable $traversable)
+    {
+        return new IteratorIterator($traversable);
     }
 }